kdijstrapath fixed

+++ functions/kdijkstra_path.py
@@ -43,7 +43,7 @@ class Function(FunctionBase):
     def getQuery(self, args):
         return """
             SELECT seq, id1 AS path, id2 AS node, id3 AS edge, cost FROM pgr_kdijkstraPath('
-                SELECT %(id)s AS id,
+                SELECT %(id)s::int4 AS id,
                     %(source)s::int4 AS source,
                     %(target)s::int4 AS target,
                     %(cost)s::float8 AS cost%(reverse_cost)s
